Add files via upload
Hey if your following along I'm skipping to 21 just because it has a large part on abstract classes and while I was learning Java data structures its best to go from concrete classes to abstract. If you look at the data structures in C and notice how its put together as if he wanted you to learn basic FIFO and LIFO  and leap too balanced trees then to advance topics like FFT. I strongly understand that is the overall best way to learn OOP. If your purely coming from python the learning curve is long, thick, and a head of a bump to get over. Although KEEP ON WORKING!