From personal computer
This is a very important chapter. Please take time to master it abstract classes is very important to OOP. Also, its what makes C++ the superset of C. Next, I will be skipping to chapter 30. If wanted I will loop back but just ask and I'll just send you info on a composite of things.